ci: Drop generic src-rpm-rebuild test
Rebuilding packages in general is (1) very expensive (and times out our
test for large packages, hello glibc), and (2) mostly unnecessary (as
they run right after we built a package in Konflux).

Move it to rootfiles, so that we still retain some coverage of our
recipe for rebuilding packages, but on a known-good/small package.