Add systemd-stub
None of systemd's services and tools are useful in our application
containers, they are just dead weight. Introduce a systemd-stub rpm
which Provides/all relevant systemd rpms.

Use a high version number "999" so that we satisfy versioned
dependencies, like `rpm`'s "Conflicts: systemd < 253.5-6", as well as
`systemd-libs`' versioned conflict with systemd-shared (which we want to
stub). Due to the latter, the stub's version number must always be
higher than the real systemd's.

Note that we do *not* `Obsoletes: systemd`. Thus an explicit
`dnf install systemd` will install the real thing.

This saves about 5 MB on e.g. the postgresql image.