ci: Use ci/builder image in create-local-repo.sh
The builder image is our reference CI place, and avoids having to
install createrepo_c.

Don't call the script with an explicit `--image`, its default `$IMAGE`
is the one we want to use, and there is no reason to try and run this in
a different one -- how create-local-repo.sh works should be an internal
implementation detail, and without the `dnf install` it's not going to
work with the core-runtime image.