dist_git: Accept and ignore local Release: modifications
We will have to rebuild some packages to clean up after the duplicate
Konflux builds/publishes. Rebuilds are a common operation in
distro-land, in case build dependencies changed significantly.

So if our only delta is in the `Release:` field of the spec, ignore that
and import a new upstream version anyway.

This is too inefficient with `filecmp`. Call `diff` instead, which has a
very handy `--ignore-matching-lines` option for exactly such cases.