ncurses, nghttp3: No-change rebuilds
These two packages were rebuilt and republished yesterday, which
resulted in overwriting the orignal rpms with a different version.

Bump their Release: to trigger a proper rebuild. Note that since commit
e2d0db38302df2 this doesn't count as local modification any more.

We don't yet have a full solution for handling `%autorelease` -- the
value of that will always be "1" for us due to our monorepo. So just
replace that with "1build1" for the time being.